# Optimize `decode_gid` with flag caching + Add tests and benchmark