--- a/pytmx/layer.py
+++ b/pytmx/layer.py
@@ -5,5 +5,4 @@ from .image_layer import TiledImageLayer
 from .object_group import TiledObjectGroup
 from .tile_layer import TiledTileLayer
 
-
 TiledLayer = Union[TiledTileLayer, TiledImageLayer, TiledGroupLayer, TiledObjectGroup]