--- a/pytmx/property.py
+++ b/pytmx/property.py
@@ -20,6 +20,7 @@ Tiled property model.
 """
 
 from xml.etree import ElementTree
+
 from .element import TiledElement